pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "fastapi>=0.115.12",
    "sqlalchemy>=2.0.40",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]
